requirements.txt
openai
langchain
python-docx